test: remove stale agent tests moved to sdk-agent package
