+++ examples/agent/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@wraith-protocol/example-stellar-spectre-agent",
+  "name": "@wraith-protocol/example-agent",
   "private": true,
   "type": "module",
-  "main": "index.ts",
+  "main": "client.ts",
   "scripts": {
-    "start": "npx tsx index.ts"
+    "start": "npx tsx client.ts"
   },
   "dependencies": {
     "@wraith-protocol/sdk": "file:../..",
