+++ packages/sdk-agent/src/index.ts
@@ -1,0 +1,17 @@
+export { Wraith, WraithAgent } from './client';
+export { Chain } from './types';
+export type {
+  WraithConfig,
+  AgentConfig,
+  AgentInfo,
+  ChatResponse,
+  ToolCall,
+  Balance,
+  Payment,
+  Invoice,
+  Schedule,
+  TxResult,
+  PrivacyReport,
+  Notification,
+  Conversation,
+} from './types';
