+++ packages/sdk-agent/test/tools.test.ts
