+++ src/agent/client-local.ts
@@ -1,0 +1,2 @@
+export class Wraith {}
+export class WraithAgent {}
