+++ src/agent/index.ts
@@ -1,5 +1,14 @@
+/**
+ * @deprecated Import from '@wraith-protocol/sdk-agent' instead. Agent functionality has been moved to a separate package.
+ */
 export { Wraith, WraithAgent } from './client';
+/**
+ * @deprecated Import from '@wraith-protocol/sdk-agent' instead. Agent functionality has been moved to a separate package.
+ */
 export { Chain } from './types';
+/**
+ * @deprecated Import from '@wraith-protocol/sdk-agent' instead. Agent functionality has been moved to a separate package.
+ */
 export type {
   WraithConfig,
   AgentConfig,
